+++ src/modules/navigator/mission_block.cpp
@@ -452,8 +452,8 @@ MissionBlock::issue_command(const mission_item_s &item)
 		vcmd.param2 = item.params[1];
 		vcmd.param3 = item.params[2];
 		vcmd.param4 = item.params[3];
-		vcmd.param5 = item.params[4];
-		vcmd.param6 = item.params[5];
+		vcmd.param5 = (double)item.params[4];
+		vcmd.param6 = (double)item.params[5];
 
 		if (item.nav_cmd == NAV_CMD_DO_SET_ROI_LOCATION && item.altitude_is_relative) {
 			vcmd.param7 = item.params[6] + _navigator->get_home_position()->alt;
@@ -644,8 +644,8 @@ MissionBlock::set_land_item(struct mission_item_s *item, bool at_current_locatio
 
 	/* use current position */
 	if (at_current_location) {
-		item->lat = NAN; //descend at current position
-		item->lon = NAN; //descend at current position
+		item->lat = (double)NAN; //descend at current position
+		item->lon = (double)NAN; //descend at current position
 		item->yaw = _navigator->get_local_position()->yaw;
 
 	} else {
